+++ common/src/caps.spec.ts
@@ -16,7 +16,7 @@ describe('CAPS', () => {
       passwordMin: 1,
       passwordMax: 100,
       groupColor: 32,
-      factoriesPerRoom: 300,
+      factoriesPerRoom: 150,
       ownedRoomsPerUser: 10,
       membershipsPerUser: 25,
       string: 10000,